Add files via upload
Add origin-traceable source file with markup links.

This file includes markup that traces the origin of the repository, based on a detailed response I made to a post on r/combatfootage around 29 May 2025. The entire history and rationale for the repository should be traceable using the contents of this file.